Implement reading from local json.

+++ api/src/main/java/com/adamstyrc/extensions/DateExtensions.kt
@@ -5,4 +5,4 @@ import java.util.*
 
 private val YEAR_FORMAT = SimpleDateFormat("yyyy", Locale.US)
 
-fun Date.year() = YEAR_FORMAT.format(this).toInt()
+fun Date.year() = YEAR_FORMAT.format(this)
